+++ games/main.ts
@@ -11,7 +11,6 @@ export * from "./legends/entities"
 export * from "./legends/Legends"
 export * from "./lobby/Lobby"
 export * from "./craft/Craft"
-export * from "./soccer/entities"
 export * from "./soccer/Soccer"
 export * from "./strike/Strike"
 export * from "./strike/StrikeSystem"
